backend/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
numpy==1.23.1
packaging==21.3
pycodestyle==2.9.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
scikit-learn==1.1.1
scipy==1.9.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.39
threadpoolctl==3.1.0
toml==0.10.2
urllib3==1.26.11
Werkzeug==2.1.2
zipp==3.8.0
